% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_valid_files.R
\name{are_valid_files}
\alias{are_valid_files}
\title{Checks if files are valid parameter files}
\usage{
are_valid_files(filenames)
}
\arguments{
\item{filenames}{the names of the file to be checked}
}
\value{
TRUE or FALSE
}
\description{
Checks if files are valid parameter files
}
\examples{
  testit::assert(
    are_valid_files(
      c(
        find_path("toy_example_1.RDa"),
        find_path("toy_example_2.RDa")
       )
     )
   )
}
\author{
Richel Bilderbeek
}
