% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_parameter_files_article_20161228.R
\name{create_parameter_files_article_20161228}
\alias{create_parameter_files_article_20161228}
\title{Creates the parameter files in the article
Now every parameter setting has its own seed}
\usage{
create_parameter_files_article_20161228()
}
\value{
The filenames of all parameter files created
}
\description{
Creates the parameter files in the article
Now every parameter setting has its own seed
}
\author{
Richel Bilderbeek
}
