% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_params.R
\name{create_params}
\alias{create_params}
\title{Create the parameters}
\usage{
create_params(speciation_initiation_rate, speciation_completion_rate,
  extinction_rate, crown_age, crown_age_sigma = 0.01, sampling_method,
  mutation_rate, sequence_length, mcmc_length, minimal_ess = 200,
  tree_sim_rng_seed, alignment_rng_seed, beast2_rnd_seed)
}
\description{
Create the parameters
}
\author{
Richel J.C. Bilderbeek
}
