% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_simulation.R
\name{do_simulation}
\alias{do_simulation}
\title{Run the full simulation pipeline on a parameter file on a local computer.}
\usage{
do_simulation(filename)
}
\arguments{
\item{filename}{name of the parameter file}
}
\value{
Nothing. It does modify the input filename
}
\description{
Run the full simulation pipeline on a parameter file on a local computer.
}
\author{
Richel Bilderbeek
}
