% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_parameters.R
\name{extract_age}
\alias{extract_age}
\title{Extract the crown age from a file}
\usage{
extract_age(file)
}
\arguments{
\item{file}{A loaded parameter file}
}
\value{
a floating point value for the crown age in time units
}
\description{
Extract the crown age from a file
}
\note{
extract_crown_age does exactly the same
}
\examples{
  file <- read_file(find_path("toy_example_1.RDa"))
  age <- extract_age(file)
  testit::assert(age == 5)
}
\author{
Richel Bilderbeek
}
