% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_parameters.R
\name{extract_fixed_crown_age}
\alias{extract_fixed_crown_age}
\title{Extract if the crown age is fixed from a file}
\usage{
extract_fixed_crown_age(file)
}
\arguments{
\item{file}{A loaded parameter file}
}
\value{
a boolean indicating if the crown age is fixed
}
\description{
Extract if the crown age is fixed from a file
}
\examples{
  file <- read_file(find_path("toy_example_1.RDa"))
  fixed_crown_age <- extract_fixed_crown_age(file)
  testit::assert(fixed_crown_age == FALSE)
}
\author{
Richel Bilderbeek
}
