% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_parameters.R
\name{extract_mutation_rate}
\alias{extract_mutation_rate}
\title{Extract the mutation rate parameter value from a file}
\usage{
extract_mutation_rate(file)
}
\arguments{
\item{file}{A loaded parameter file}
}
\value{
the mutation rate
}
\description{
Extract the mutation rate parameter value from a file
}
\examples{
  file <- read_file(find_path("toy_example_1.RDa"))
  mutation_rate <- extract_mutation_rate(file)
  testit::assert(mutation_rate == 0.01)
}
\author{
Richel Bilderbeek
}
