% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_parameters.R
\name{extract_parameters}
\alias{extract_parameters}
\title{Extract the parameters from a file created by wiritttes::read_file}
\usage{
extract_parameters(file)
}
\arguments{
\item{file}{A loaded parameter file}
}
\value{
the value of the ERG parameter
}
\description{
Extract the parameters from a file created by wiritttes::read_file
}
\examples{
  file <- read_file(find_path("toy_example_1.RDa"))
  parameters <- extract_parameters(file)
  testit::assert(nrow(parameters) > 1)
  testit::assert(ncol(parameters) > 1)
}
\author{
Richel Bilderbeek
}

