% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_parameters.R
\name{extract_seed}
\alias{extract_seed}
\title{Extract the RNG seed parameter value from a file}
\usage{
extract_seed(file)
}
\arguments{
\item{file}{A loaded parameter file}
}
\value{
the RNG seed rate
}
\description{
Extract the RNG seed parameter value from a file
}
\examples{
  file <- read_file(find_path("toy_example_1.RDa"))
  rng_seed <- extract_seed(file)
  testit::assert(rng_seed == 1)
}
\author{
Richel Bilderbeek
}
