% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pbd_output.R
\name{get_pbd_output}
\alias{get_pbd_output}
\title{Get the PBD simulation output}
\usage{
get_pbd_output(file)
}
\arguments{
\item{file}{A loaded parameter file}
}
\value{
the PBD simulation output
}
\description{
Get the PBD simulation output
}
\examples{
  filename <- wiritttes::find_path("toy_example_1.RDa")
  file <- wiritttes::read_file(filename)
  pbd_output <- wiritttes::get_pbd_output(file)
  testit::assert(ribir::is_pbd_sim_output(pbd_output))
}
\author{
Richel Bilderbeek
}
