% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_posterior_tree.R
\name{get_posterior_tree}
\alias{get_posterior_tree}
\title{Extract a BEAST2 posterior tree from a file}
\usage{
get_posterior_tree(file, sti, ai, pi, si)
}
\arguments{
\item{file}{A loaded parameter file}

\item{sti}{the species tree index, a value from 1 to and including 2}

\item{ai}{the alignment index,
the ai-th alignment of that species tree,
a value from 1 to and including
the number of alignments per species tree}

\item{pi}{the posterior index,
the pi-th posterior of that alignment,
a value from 1 to and including
the number of posteriors per alignment}

\item{si}{the posterior state index,
the si-th state in the posterior,
a value from 1 to and including
the number of posterior states per posterior}
}
\value{
the posterior
}
\description{
Extract a BEAST2 posterior tree from a file
}
\examples{
  # Read a file with one or more posteriors
  file <- read_file(find_path("toy_example_1.RDa"))
  # Pick the indices of the posterior to extract
  sti  <- 1 # Species tree index
  ai   <- 1 # Alignment index
  pi   <- 1 # Posterior index
  si   <- 1 # Posterior state index
  # Extract the posterior
  tree <- get_posterior_tree(file = file, sti = sti, ai = 1, pi = 1, si = 1)
  # Check that it is indeed a phylogeny
  testit::assert(ribir::is_phylogeny(tree))
}
\author{
Richel Bilderbeek
}
