% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_posteriors.R
\name{get_posteriors}
\alias{get_posteriors}
\title{Extract the BEAST2 posterior phyogenies
  from a file}
\usage{
get_posteriors(file)
}
\arguments{
\item{file}{A loaded parameter file}
}
\value{
all posteriors
}
\description{
Extract the BEAST2 posterior phyogenies
  from a file
}
\examples{
  filename <- wiritttes::find_path("toy_example_1.RDa")
  file <- wiritttes::read_file(filename)
  posteriors <- wiritttes::get_posteriors(file)
  testit::assert(length(posteriors) == 2)
  testit::assert("trees" \%in\% names(posteriors[[1]][[1]]))
  testit::assert("estimates" \%in\% names(posteriors[[1]][[1]]))
  testit::assert("likelihood" \%in\% names(posteriors[[1]][[1]]$estimates))
}
\author{
Richel Bilderbeek
}
