% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species_tree.R
\name{get_species_tree_oldest}
\alias{get_species_tree_oldest}
\title{Get the species tree with oldest subspecies representing each good species}
\usage{
get_species_tree_oldest(file)
}
\arguments{
\item{file}{A loaded parameter file}
}
\value{
the species_tree
}
\description{
Get the species tree with oldest subspecies representing each good species
}
\author{
Richel Bilderbeek
}
