% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_valid_file.R
\name{is_valid_file}
\alias{is_valid_file}
\title{Checks if one or more files are valid parameter files}
\usage{
is_valid_file(filename)
}
\arguments{
\item{filename}{the name of the file to be checked}
}
\value{
TRUE or FALSE
}
\description{
Checks if one or more files are valid parameter files
}
\examples{
  testit::assert(is_valid_file(find_path("toy_example_1.RDa")))
}
\author{
Richel Bilderbeek
}
