% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_files}
\alias{read_files}
\title{Read multiple parameter (.RDa) files}
\usage{
read_files(filenames)
}
\arguments{
\item{filenames}{names of the parameter (.RDa) file}
}
\value{
the file as R data
}
\description{
Read multiple parameter (.RDa) files
}
\examples{
  files <- read_files(
    find_paths(c("toy_example_1.RDa", "toy_example_2.RDa")))
  testit::assert(extract_sirg(files[[1]]) > 0.0)
  testit::assert(extract_siri(files[[2]]) > 0.0)
}
\author{
Richel Bilderbeek
}

