% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pbd_output.R
\name{get_pbd_output}
\alias{get_pbd_output}
\title{Get the PBD simulation output}
\usage{
get_pbd_output(file)
}
\arguments{
\item{file}{A loaded parameter file}
}
\value{
the PBD simulation output
}
\description{
Get the PBD simulation output
}
\author{
Richel Bilderbeek
}
